backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.102.0
anyio==4.13.0
ast_serialize==0.3.0
asyncpg==0.31.0
attrs==26.1.0
bcrypt==4.1.3
beautifulsoup4==4.14.3
billiard==4.2.4
black==26.3.1
boto3==1.43.6
botocore==1.43.6
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==48.0.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.18.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.29.0
flake8==7.3.0
frozenlist==1.8.0
fsspec==2026.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.196.0
google-auth==2.52.0
google-auth-httplib2==0.4.0
google-genai==2.0.1
google-generativeai==0.8.6
googleapis-common-protos==1.75.0
greenlet==3.5.0
h11==0.16.0
hf-xet==1.5.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.14.0
idna==3.14
importlib_metadata==9.0.0
iniconfig==2.3.0
isort==8.0.1
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
jq==1.11.0
json_repair==0.59.9
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
langchain==1.3.0
langchain-classic==1.0.7
langchain-core==1.4.0
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langgraph==1.2.0
langgraph-checkpoint==4.1.0
langgraph-prebuilt==1.1.0
langgraph-sdk==0.3.14
langsmith==0.8.4
lark==1.3.1
librt==0.11.0
litellm==1.80.0
lxml==6.1.0
markdown-it-py==4.2.0
markdownify==0.13.1
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
multidict==6.7.1
mypy==2.0.0
mypy_extensions==1.1.0
numpy==1.26.3
oauthlib==3.3.1
openai==1.99.9
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
pandas==2.3.3
passlib==1.7.4
pathspec==1.1.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.5.2
proto-plus==1.28.0
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.13.4
pydantic_core==2.46.4
pyflakes==3.4.0
Pygments==2.20.0
PyJWT==2.12.1
pymongo==4.5.0
pyparsing==3.3.2
pypdfium2==5.8.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-jobspy==1.1.82
python-jose==3.5.0
python-multipart==0.0.28
pytokens==0.4.1
pytz==2026.2
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
regex==2024.11.6
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.17.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==0.37.2
stripe==15.1.0
tenacity==9.1.4
tiktoken==0.12.0
tls-client==1.0.1
tokenizers==0.23.1
tqdm==4.67.3
typer==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.7.0
uuid_utils==0.15.0
uvicorn==0.25.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.7.0
websockets==16.0
xxhash==3.7.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
